Flutter/flutter_project/google_auth_helper/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
google-auth>=2.35.0
python-dotenv>=1.0.1
